Fitting function to be called by user

This function creates a list of parameters, sets up TMB object and attempts to do fitting / estimation

fit(
  data_list,
  silent = FALSE,
  inits = NULL,
  control = list(eval.max = 2000, iter.max = 1000, rel.tol = 1e-10),
  limits = FALSE
)

Arguments

data_list

A list of data, as output from create_data

silent

Boolean passed to TMB::MakeADFun, whether to be verbose or not (defaults to FALSE)

inits

Optional named list of parameters for starting values, defaults to NULL

control

Optional control list for stats::nlminb. For arguments see ?nlminb. Defaults to eval.max=2000, iter.max=1000, rel.tol=1e-10. For final model runs, the rel.tol should be even smaller

limits

Whether to include limits for stats::nlminb, defaults to FALSE

Examples

data(fishdist)

# example of fitting fixed effects, no trends, no random effects
set.seed(1)
datalist <- create_data(fishdist[which(fishdist$year > 1970), ],
  asymmetric_model = FALSE,
  est_mu_re = FALSE, est_sigma_re = FALSE
)
fit <- fit(datalist)
#> Warning: Package version inconsistency detected.
#> TMB was built with Matrix version 1.3.4
#> Current Matrix version is 1.4.0
#> Please re-install 'TMB' from source using install.packages('TMB', type = 'source') or ask CRAN for a binary version of 'TMB' matching CRAN's 'Matrix' package
#> Order of parameters:
#>  [1] "theta"             "b_mu"              "log_sigma_mu_devs"
#>  [4] "mu_devs"           "b_sig1"            "b_sig2"           
#>  [7] "log_sigma1"        "sigma1_devs"       "log_sigma2"       
#> [10] "sigma2_devs"       "log_obs_sigma"     "log_tdf_1"        
#> [13] "log_tdf_2"         "log_beta_1"        "log_beta_2"       
#> Not matching template order:
#>  [1] "log_sigma1"        "sigma1_devs"       "log_sigma2"       
#>  [4] "sigma2_devs"       "theta"             "mu_devs"          
#>  [7] "log_sigma_mu_devs" "log_tdf_1"         "log_tdf_2"        
#> [10] "log_beta_1"        "log_beta_2"        "log_obs_sigma"    
#> [13] "b_mu"              "b_sig1"            "b_sig2"           
#> Your parameter list has been re-ordered.
#> (Disable this warning with checkParameterOrder=FALSE)
#> Constructing atomic qnorm1
#> outer mgc:  4616.648 
#> outer mgc:  200.2752 
#> outer mgc:  101.9723 
#> outer mgc:  23.30524 
#> outer mgc:  15.92007 
#> outer mgc:  107.3552 
#> outer mgc:  43.77026 
#> outer mgc:  59.58926 
#> outer mgc:  83.53679 
#> outer mgc:  66.19899 
#> outer mgc:  12.06982 
#> outer mgc:  7.36865 
#> outer mgc:  30.63229 
#> outer mgc:  11.6692 
#> outer mgc:  18.54689 
#> outer mgc:  9.998038 
#> outer mgc:  14.13124 
#> outer mgc:  8.587495 
#> outer mgc:  2.809278 
#> outer mgc:  1.058406 
#> outer mgc:  0.5065038 
#> outer mgc:  1.522275 
#> outer mgc:  0.4245952 
#> outer mgc:  1.187253 
#> outer mgc:  1.419662 
#> outer mgc:  1.554793 
#> outer mgc:  0.1749995 
#> outer mgc:  0.8611607 
#> outer mgc:  0.1031742 
#> outer mgc:  0.1586227 
#> outer mgc:  0.0492994 
#> outer mgc:  0.08334353 
#> outer mgc:  0.02874414 
#> outer mgc:  0.03044152 
#> outer mgc:  0.02443265 
#> outer mgc:  0.009828507 
#> outer mgc:  0.0005049515 
#> outer mgc:  0.0005049515 
#> outer mgc:  0.03362425 
#> outer mgc:  0.03463415 
#> outer mgc:  0.03299622 
#> outer mgc:  0.03235756 
#> outer mgc:  0.03335977 
#> outer mgc:  0.03344632 
#> outer mgc:  0.03563269 
#> outer mgc:  0.03553032 
#> outer mgc:  0.03370598 
#> outer mgc:  0.03310011 
#> outer mgc:  0.03630051 
#> outer mgc:  0.03631481 
#> outer mgc:  0.03549645 
#> outer mgc:  0.03566656 
#> outer mgc:  0.04009836 
#> outer mgc:  0.03977849 
#> outer mgc:  0.03472723 
#> outer mgc:  0.03498348 
#> outer mgc:  0.03618135 
#> outer mgc:  0.03643397 
#> outer mgc:  0.9685688 
#> outer mgc:  0.9714343 
#> outer mgc:  0.00880458 
#> outer mgc:  0.009232193 
#> outer mgc:  0.03053612 
#> outer mgc:  0.03099023 
#> outer mgc:  1.34898 
#
# # example of model with random effects in means only, and symmetric distribution
# set.seed(1)
# datalist <- create_data(fishdist[which(fishdist$year > 1970),], asymmetric_model = FALSE,
#                      est_sigma_re = FALSE)
# fit <- fit(datalist)
# # example of model with random effects in variances
# set.seed(1)
# datalist <- create_data(fishdist[which(fishdist$year > 1970),], asymmetric_model = TRUE,
#                          est_mu_re = TRUE)
# fit <- fit(datalist)
#
# # example of model with poisson response
# set.seed(1)
# datalist <- create_data(fishdist[which(fishdist$year > 1970),], asymmetric_model = FALSE,
#                         est_sigma_trend=FALSE, est_mu_trend=FALSE, est_mu_re = TRUE,
#                         family="poisson")
# fit <- fit(datalist)