Generate metadata

Generate metadata for OpenGWAS datasets.

generate_metadata(
  munged_files,
  infer_builds = TRUE,
  drop_na_N = TRUE,
  trait_groups = NULL,
  force_new = FALSE,
  save_dir = tempdir(),
  verbose = TRUE
)

Arguments

munged_files
List of paths to summary statostics files that have been munged with import_sumstats.

infer_builds
Infer genome builds of each file.

drop_na_N
Drop rows without "N" (total sample size).

trait_groups
[Optional] A vector of strings. Creates a new column "trait_groups" based on a series of substring searches.

force_new
Create a new metadata file even when one already exists in save_dir.

save_dir
Directory to save metadata to.

verbose
Print messages.

Value

data.frame of OpenGWAS metadata.