Map SNPs to genes: txdb

Map both coding and non-coding SNPs to genes using #' TxDb.Hsapiens.UCSC.hg19.knownGene or TxDb.Hsapiens.UCSC.hg38.knownGene as a reference.

map_snps_txdb(
  sumstats,
  promoter_upstream = 35000,
  promoter_downstream = 10000,
  queries = c("promoter", "coding", "intron", "threeUTR", "splicesite"),
  all_variants = FALSE,
  ref_genome = "GRCh37",
  return_merged = TRUE,
  nCores = 1,
  verbose = TRUE
)

Arguments

sumstats
GWAS summary statistics munged by format_sumstats.

promoter_upstream
How many basepairs upstream of known promoters to search for genes.

promoter_downstream
How many basepairs dowmstream of known promoters to search for genes.

queries
Genomic region types to search for overlap with gr. See AllVariants for details.

all_variants
Search all genomic regions regardless of type. This can take longer and returns less targeted results.

ref_genome
name of the reference genome used for the GWAS ("GRCh37" or "GRCh38"). Argument is case-insensitive. Default is NULL which infers the reference genome from the data.

verbose
Print messages.

Value

hits data.table

Details

Permits selection of different region types with queries argument.