Fitting function to be called by user

This function creates a list of parameters, sets up TMB object and attempts to do fitting / estimation

fit(
  data_list,
  silent = FALSE,
  inits = NULL,
  control = list(eval.max = 2000, iter.max = 1000, rel.tol = 1e-10),
  limits = NULL,
  fit_model = TRUE
)

Arguments

data_list

A list of data, as output from create_data

silent

Boolean passed to TMB::MakeADFun, whether to be verbose or not (defaults to FALSE)

inits

Optional named list of parameters for starting values, defaults to NULL

control

Optional control list for stats::nlminb. For arguments see ?nlminb. Defaults to eval.max=2000, iter.max=1000, rel.tol=1e-10. For final model runs, the rel.tol should be even smaller

limits

Whether to include limits for stats::nlminb. Can be a list of (lower, upper), or TRUE to use suggested hardcoded limits. Defaults to NULL, where no limits used.

fit_model

Whether to fit the model. If not, returns a list including the data, parameters, and initial values. Defaults to TRUE

Examples

data(fishdist)

# example of fitting fixed effects, no trends, no random effects
set.seed(1)
datalist <- create_data(fishdist[which(fishdist$year > 1970), ],
  asymmetric_model = FALSE,
  est_mu_re = FALSE, est_sigma_re = FALSE
)
fit <- fit(datalist)
#> Warning: Package version inconsistency detected.
#> TMB was built with Matrix version 1.4.1
#> Current Matrix version is 1.4.0
#> Please re-install 'TMB' from source using install.packages('TMB', type = 'source') or ask CRAN for a binary version of 'TMB' matching CRAN's 'Matrix' package
#> Order of parameters:
#>  [1] "theta"             "b_mu"              "log_sigma_mu_devs"
#>  [4] "mu_devs"           "b_sig1"            "b_sig2"           
#>  [7] "log_sigma1_sd"     "sigma1_devs"       "log_sigma2_sd"    
#> [10] "sigma2_devs"       "log_obs_sigma"     "log_tdf_1"        
#> [13] "log_tdf_2"         "log_beta_1"        "log_beta_2"       
#> Not matching template order:
#>  [1] "log_sigma1_sd"     "sigma1_devs"       "log_sigma2_sd"    
#>  [4] "sigma2_devs"       "theta"             "mu_devs"          
#>  [7] "log_sigma_mu_devs" "log_tdf_1"         "log_tdf_2"        
#> [10] "log_beta_1"        "log_beta_2"        "log_obs_sigma"    
#> [13] "b_mu"              "b_sig1"            "b_sig2"           
#> Your parameter list has been re-ordered.
#> (Disable this warning with checkParameterOrder=FALSE)
#> Constructing atomic qnorm1
#> outer mgc:  25077630 
#> outer mgc:  414951.8 
#> outer mgc:  238621.3 
#> outer mgc:  33074.13 
#> outer mgc:  20763.32 
#> outer mgc:  3993.586 
#> outer mgc:  2584.106 
#> outer mgc:  625.1119 
#> outer mgc:  469.8424 
#> outer mgc:  286.8392 
#> outer mgc:  203.295 
#> outer mgc:  56.71973 
#> outer mgc:  171.7801 
#> outer mgc:  196.9294 
#> outer mgc:  195.0662 
#> outer mgc:  224.446 
#> outer mgc:  194.0984 
#> outer mgc:  184.3113 
#> outer mgc:  266.782 
#> outer mgc:  354.6835 
#> outer mgc:  205.8583 
#> outer mgc:  110.8358 
#> outer mgc:  134.4663 
#> outer mgc:  58.1291 
#> outer mgc:  78.33889 
#> outer mgc:  25.10897 
#> outer mgc:  16.25894 
#> outer mgc:  70.73127 
#> outer mgc:  6.796327 
#> outer mgc:  34.34238 
#> outer mgc:  7.059354 
#> outer mgc:  14.86802 
#> outer mgc:  3.320803 
#> outer mgc:  3.809977 
#> outer mgc:  2.35201 
#> outer mgc:  11.02337 
#> outer mgc:  1.78548 
#> outer mgc:  1.48053 
#> outer mgc:  2.347749 
#> outer mgc:  5.902279 
#> outer mgc:  1.237691 
#> outer mgc:  2.321378 
#> outer mgc:  0.9762488 
#> outer mgc:  1.708979 
#> outer mgc:  0.6839345 
#> outer mgc:  0.172727 
#> outer mgc:  0.7626058 
#> outer mgc:  0.7035818 
#> outer mgc:  0.05360936 
#> outer mgc:  0.1591468 
#> outer mgc:  0.01323017 
#> outer mgc:  0.005855694 
#> outer mgc:  0.007077789 
#> outer mgc:  0.0002211938 
#> outer mgc:  4.755575e-05 
#> outer mgc:  4.755575e-05 
#> outer mgc:  0.03413568 
#> outer mgc:  0.03412265 
#> outer mgc:  0.03268433 
#> outer mgc:  0.03266939 
#> outer mgc:  0.0333956 
#> outer mgc:  0.03341042 
#> outer mgc:  0.03559128 
#> outer mgc:  0.03557166 
#> outer mgc:  0.03341511 
#> outer mgc:  0.03339091 
#> outer mgc:  0.03628597 
#> outer mgc:  0.03632927 
#> outer mgc:  0.03558214 
#> outer mgc:  0.03558079 
#> outer mgc:  0.03993684 
#> outer mgc:  0.03993993 
#> outer mgc:  0.03487398 
#> outer mgc:  0.03483665 
#> outer mgc:  0.03632153 
#> outer mgc:  0.03629371 
#> outer mgc:  0.9690781 
#> outer mgc:  0.970923 
#> outer mgc:  0.009047906 
#> outer mgc:  0.00898857 
#> outer mgc:  0.03077942 
#> outer mgc:  0.03074658 
#> outer mgc:  1.34898 
#
# # example of model with random effects in means only, and symmetric distribution
# set.seed(1)
# datalist <- create_data(fishdist[which(fishdist$year > 1970),], asymmetric_model = FALSE,
#                      est_sigma_re = FALSE)
# fit <- fit(datalist)
# # example of model with random effects in variances
# set.seed(1)
# datalist <- create_data(fishdist[which(fishdist$year > 1970),], asymmetric_model = TRUE,
#                          est_mu_re = TRUE)
# fit <- fit(datalist)
#
# # example of model with poisson response
# set.seed(1)
# datalist <- create_data(fishdist[which(fishdist$year > 1970),], asymmetric_model = FALSE,
#                         est_sigma_trend=FALSE, est_mu_trend=FALSE, est_mu_re = TRUE,
#                         family="poisson")
# fit <- fit(datalist)